Optimized .next deletion

+++ index.js
@@ -70,7 +70,9 @@ async function installRelativeDepsWithNext() {
   if (reloaded) {
     console.log(`[relative-deps] Reloading next dev evironment`)
     if (abortSignal) abortSignal.abort()
-    rimraf.sync(path.join(process.cwd(), ".next"))
+    if (fs.existsSync(".next")) {
+      rimraf.sync(path.join(process.cwd(), ".next"))
+    }
     abortSignal = new AbortController()
     spawn.sync(["run", "dev"], { cwd: process.cwd(), stdio: [0, 1, 2], signal: abortSignal })
     console.log(`[relative-deps] Reloading next dev evironment... DONE`)
